+++ examples/src/main/java/com/wit/android/support/fragment/examples/fragment/FragmentsFactory.java
@@ -23,6 +23,7 @@ package com.wit.android.support.fragment.examples.fragment;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 
+import com.wit.android.support.fragment.annotation.FactoryFragment;
 import com.wit.android.support.fragment.annotation.FactoryFragments;
 import com.wit.android.support.fragment.manage.BaseFragmentFactory;
 import com.wit.android.support.fragment.manage.FragmentController;
@@ -38,9 +39,7 @@ import com.wit.android.support.fragment.manage.FragmentTransition;
 @FactoryFragments({
 		FragmentsFactory.TRANSITIONS,
 		FragmentsFactory.LIST,
-		FragmentsFactory.GRID,
-		FragmentsFactory.ACTION_BAR,
-		FragmentsFactory.ANNOTATIONS
+		FragmentsFactory.GRID
 })
 public class FragmentsFactory extends BaseFragmentFactory {
 
@@ -49,29 +48,16 @@ public class FragmentsFactory extends BaseFragmentFactory {
 	 */
 	//private static final String TAG = FragmentsFactory.class.getSimpleName();
 
-	/**
-	 *
-	 */
 	public static final int TRANSITIONS = 0x01;
 
-	/**
-	 *
-	 */
 	public static final int LIST = 0x02;
 
-	/**
-	 *
-	 */
 	public static final int GRID = 0x03;
 
-	/**
-	 *
-	 */
+	@FactoryFragment
 	public static final int ACTION_BAR = 0x04;
 
-	/**
-	 *
-	 */
+	@FactoryFragment
 	public static final int ANNOTATIONS = 0x05;
 
 	/**
@@ -108,7 +94,7 @@ public class FragmentsFactory extends BaseFragmentFactory {
 			case GRID:
 				return GridFragmentImpl.newInstance();
 			case ACTION_BAR:
-				return ActionBarFragmentImpl.newInstance();
+				return ActionBarFragmentImpl.newInstance(params);
 			case ANNOTATIONS:
 				return AnnotatedFragment.newInstance();
 		}